fix: tag model renamed file
