+++ packages/contracts/src/candidate.model.ts
@@ -12,12 +12,12 @@ import { ICandidateDocument } from './candidate-document.model';
 import { IOrganizationDepartment } from './organization-department.model';
 import { IOrganizationPosition } from './organization-positions.model';
 import { IContact } from './contact.model';
-import { ITag } from './tag-entity.model';
+import { ITag } from './tag.model';
 import { IEmployee, PayPeriodEnum } from './employee.model';
 
 export interface IRelationalCandidate {
-    readonly candidate?: ICandidate;
-    readonly candidateId?: ICandidate['id'];
+	readonly candidate?: ICandidate;
+	readonly candidateId?: ICandidate['id'];
 }
 
 export interface ICandidate extends IBasePerTenantAndOrganizationEntityModel {
@@ -118,4 +118,3 @@ export interface ICandidateViewModel extends IBasePerTenantAndOrganizationEntity
 	isArchived?: boolean;
 	status?: CandidateStatusEnum;
 }
-
