+++ packages/contracts/src/entity-with-members.model.ts
@@ -1,6 +1,6 @@
 import { IEmployee } from './employee.model';
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
-import { ITag } from './tag-entity.model';
+import { ITag } from './tag.model';
 
 export interface IBaseEntityWithMembers
 	extends IBasePerTenantAndOrganizationEntityModel {
