+++ packages/contracts/src/event-type.model.ts
@@ -1,6 +1,6 @@
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
 import { IEmployee, IEmployeeFindInput } from './employee.model';
-import { ITag } from './tag-entity.model';
+import { ITag } from './tag.model';
 
 export interface IEventType extends IBasePerTenantAndOrganizationEntityModel {
 	title: string;
@@ -58,5 +58,5 @@ export interface IEventTypeViewModel {
 	tags: ITag[];
 	isDefault?: boolean;
 	employeeId?: string;
-    employee?: IEmployee;
+	employee?: IEmployee;
 }
