+++ packages/contracts/src/expense-category.model.ts
@@ -1,6 +1,6 @@
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
 import { IExpense } from './expense.model';
-import { ITag } from './tag-entity.model';
+import { ITag } from './tag.model';
 
 export interface IExpenseCategory extends IBasePerTenantAndOrganizationEntityModel {
 	name: string;
