+++ packages/contracts/src/index.ts
@@ -96,7 +96,7 @@ export * from './role.model';
 export * from './seed.model';
 export * from './skill-entity.model';
 export * from './sms.model';
-export * from './tag-entity.model';
+export * from './tag.model';
 export * from './task-priority.model';
 export * from './task-size.model';
 export * from './task-status.model';
