+++ packages/contracts/src/organization-department.model.ts
@@ -1,6 +1,6 @@
 import { IBaseEntityWithMembers } from './entity-with-members.model';
 import { IEmployee } from './employee.model';
-import { ITag } from './tag-entity.model';
+import { ITag } from './tag.model';
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
 import { ICandidate } from './candidate.model';
