+++ packages/contracts/src/task.model.ts
@@ -7,7 +7,7 @@ import { IInvoiceItem } from './invoice-item.model';
 import { IOrganizationProject } from './organization-projects.model';
 import { IOrganizationSprint } from './organization-sprint.model';
 import { IOrganizationTeam } from './organization-team-model';
-import { ITag } from './tag-entity.model';
+import { ITag } from './tag.model';
 import { IUser } from './user.model';
 import { TaskStatusEnum } from './task-status.model';
 import { TaskPriorityEnum } from './task-priority.model';
